Update with dark-mode support
Also updates service worker setup and various styles
