+++ .gitignore
@@ -66,3 +66,6 @@ typings/
 
 # next.js build output
 .next
+
+# Local Netlify folder
+.netlify
