+++ css/index.css
@@ -8,6 +8,7 @@
 @import url("https://cdn.kernvalley.us/css/core-css/fonts.css");
 @import url("https://cdn.kernvalley.us/css/animate.css/animate.css");
 @import url("https://cdn.kernvalley.us/css/core-css/animations.css");
+@import url("https://cdn.kernvalley.us/css/core-css/layout/shared.css");
 @import url("./layout.css");
 
 input[hidden] + label[for].btn-toggle {
