+++ js/index.js
@@ -1,7 +1,7 @@
 import 'https://cdn.kernvalley.us/js/std-js/deprefixer.js';
 import 'https://cdn.kernvalley.us/js/std-js/shims.js';
-import './share-button.js';
-import './current-year.js';
+import 'https://cdn.kernvalley.us/components/share-button.js';
+import 'https://cdn.kernvalley.us/components/current-year.js';
 import './rental-properties.js';
 // import './gravatar-img.js';
 // import 'https://cdn.kernvalley.us/components/bacon-ipsum.js';
