+++ package.json
@@ -38,7 +38,7 @@
   "scripts": {
     "preinstall": "git submodule update --recursive --init",
     "postinstall": "npm run build:icons",
-    "start": "http-server ${npm_package_config_serve_path} -p ${npm_package_config_serve_port} -a ${npm_package_config_serve_domain} -o",
+    "start": "npm run netlify:serve",
     "test": "npm run lint:js && npm run lint:css",
     "fix": "npm run fix:js && npm run fix:css",
     "build": "npm run build:icons",
@@ -46,7 +46,10 @@
     "lint:css": "stylelint ${npm_package_config_dir_css}",
     "lint:js": "eslint ${npm_package_config_dir_js}",
     "fix:css": "stylelint ${npm_package_config_dir_css} --fix",
-    "fix:js": "eslint ${npm_package_config_dir_js} --fix"
+    "fix:js": "eslint ${npm_package_config_dir_js} --fix",
+    "netlify:deploy": "netlify deploy",
+    "netlify:serve": "netlify dev",
+    "netlify:live": "netlify dev --live"
   },
   "dependencies": {
     "svg-sprite-generator": "0.0.7"
@@ -57,6 +60,7 @@
     "eslint-plugin-async-await": "0.0.0",
     "eslint-plugin-babel": "^5.3.0",
     "http-server": "^0.11.1",
+    "netlify-cli": "^2.19.4",
     "stylelint": "^10.0.1",
     "stylelint-config-recommended": "^2.2.0",
     "stylelint-config-standard": "^18.3.0",
