+++ service-worker.js
@@ -1,103 +1,63 @@
 'use strict';
+/*eslint no-undef: 0*/
+/* 2019-10-25T13:04 */
+self.importScripts('/sw-config.js');
 
-const config = {
-	version: location.hostname === 'localhost' ? new Date().toISOString() : '1.0.0-a11',
-	stale: [
-		'/',
-		'/js/index.js',
-		'/js/share-button.js',
-		'/js/share-config.js',
-		'/js/gravatar-img.js',
-		'/js/current-year.js',
-		'/js/imgur-img.js',
-		'/js/rental-properties.js',
-		'/js/rental-property.js',
-		'/js/schema-postal-address.js',
-		'/js/slot-helpers.js',
-		'https://cdn.kernvalley.us/js/std-js/deprefixer.js',
-		'https://cdn.kernvalley.us/js/std-js/shims.js',
-		'https://cdn.kernvalley.us/js/std-js/md5.js',
-		'https://cdn.kernvalley.us/js/std-js/Notification.js',
-		'https://cdn.kernvalley.us/js/std-js/webShareApi.js',
-		'https://cdn.kernvalley.us/js/std-js/esQuery.js',
-		'https://cdn.kernvalley.us/js/std-js/functions.js',
-		'https://cdn.kernvalley.us/components/login-button.js',
-		'https://cdn.kernvalley.us/components/logout-button.js',
-		'https://cdn.kernvalley.us/components/register-button.js',
-		'https://cdn.kernvalley.us/components/gravatar-img.js',
-		'https://cdn.kernvalley.us/js/std-js/asyncDialog.js',
-		'https://cdn.kernvalley.us/js/User.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentRequest.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentRequestUpdateEvent.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentAddress.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentResponse.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/BasicCardResponse.js',
-		'https://cdn.kernvalley.us/js/PaymentAPI/BillingAddress.js',
-		'https://cdn.kernvalley.us/components/payment-form/payment-form.js',
-		'https://cdn.kernvalley.us/components/login-form/login-form.js',
-		'https://cdn.kernvalley.us/components/registration-form/registration-form.js',
-		'https://cdn.kernvalley.us/components/login-form/login-form.html',
-		'https://cdn.kernvalley.us/components/registration-form/registration-form.html',
-		'/css/index.css',
-		'/css/vars.css',
-		'/css/layout.css',
-		'/css/header.css',
-		'/css/nav.css',
-		'/css/main.css',
-		'/css/sidebar.css',
-		'/css/footer.css',
-		'https://cdn.kernvalley.us/css/core-css/rem.css',
-		'https://cdn.kernvalley.us/css/core-css/viewport.css',
-		'https://cdn.kernvalley.us/css/core-css/element.css',
-		'https://cdn.kernvalley.us/css/core-css/class-rules.css',
-		'https://cdn.kernvalley.us/css/core-css/utility.css',
-		'https://cdn.kernvalley.us/css/core-css/fonts.css',
-		'https://cdn.kernvalley.us/css/core-css/animations.css',
-		'https://cdn.kernvalley.us/css/normalize/normalize.css',
-		'https://cdn.kernvalley.us/css/animate.css/animate.css',
-		'/img/apple-touch-icon.png',
-		'/img/icon-192.png',
-		'/img/favicon.svg',
-		'/img/octicons/mail.svg',
-		'/img/logos/instagram.svg',
-		'/img/icons.svg',
-		'/img/listings/05c6cbb6-713f-486d-a867-e3f70d967bb5.jpg',
-		'https://cdn.kernvalley.us/fonts/roboto.woff2',
-		'https://cdn.kernvalley.us/fonts/Libertine.woff',
-		'/listings.json',
-	].map(path => new URL(path, location.origin).href),
-};
+self.addEventListener('install', async event => {
+	event.waitUntil((async () => {
+		try {
+			for (const key of await caches.keys()) {
+				await caches.delete(key);
+			}
 
-self.addEventListener('install', async () => {
-	const cache = await caches.open(config.version);
-	const keys = await caches.keys();
-	const old = keys.filter(k => k !== config.version);
-	await Promise.all(old.map(key => caches.delete(key)));
-
-	try {
-		await cache.addAll(config.stale);
-	} catch (err) {
-		console.error(err);
-	}
-
-	skipWaiting();
+			const cache = await caches.open(config.version);
+			await cache.addAll(config.stale);
+		} catch (err) {
+			console.error(err);
+		}
+	})());
 });
 
-self.addEventListener('activate', event => {
-	event.waitUntil(async function() {
-		clients.claim();
-	}());
-});
+self.addEventListener('activate', event => event.waitUntil(clients.claim()));
 
 self.addEventListener('fetch', event => {
-	if (event.request.method === 'GET' && config.stale.includes(event.request.url)) {
+	if (event.request.method === 'GET' && event.request.url.startsWith(location.origin)) {
 		event.respondWith((async () => {
-			const cached = await caches.match(event.request);
-			if (cached instanceof Response) {
-				return cached;
+			const url = new URL(event.request.url);
+			url.hash = '';
+
+			if (Array.isArray(config.stale) && config.stale.includes(url.href)) {
+				const cached = await caches.match(url);
+				if (cached instanceof Response) {
+					return cached;
+				}
+			} else if (Array.isArray(config.fresh) && config.fresh.includes(url.href)) {
+				if (navigator.onLine) {
+					const resp = await fetch(url.href);
+					const cache = await caches.open(config.version);
+
+					if (resp.ok) {
+						cache.add(resp.clone());
+					}
+					return resp;
+				} else {
+					return caches.match(event.request);
+				}
+			} else if (Array.isArray(config.allowed) && config.allowed.some(host => new URL(event.request.url).host === host)) {
+				const resp = await caches.match(event.request);
+				if (resp instanceof Response) {
+					return resp;
+				} else if (navigator.onLine) {
+					const resp = await fetch(event.request);
+					const cache = await caches.open(config.version);
+					cache.add(resp.clone());
+					return resp;
+				}
 			} else {
-				return await fetch(event.request);
+				return fetch(event.request.url);
 			}
 		})());
 	}
 });
+
+self.addEventListener('error', console.error);
