+++ sw-config.js
@@ -1,0 +1,74 @@
+'use strict';
+/* eslint no-unused-vars: 0 */
+const config = {
+	version: '1.0.0-a12',
+	stale: [
+		'/',
+		'/js/index.js',
+		'https://cdn.kernvalley.us/components/share-button.js',
+		'https://cdn.kernvalley.us/js/std-js/share-config.js',
+		'/js/gravatar-img.js',
+		'https://cdn.kernvalley.us/components/current-year.js',
+		'/js/imgur-img.js',
+		'/js/rental-properties.js',
+		'/js/rental-property.js',
+		'/js/schema-postal-address.js',
+		'/js/slot-helpers.js',
+		'https://cdn.kernvalley.us/js/std-js/deprefixer.js',
+		'https://cdn.kernvalley.us/js/std-js/shims.js',
+		'https://cdn.kernvalley.us/js/std-js/md5.js',
+		'https://cdn.kernvalley.us/js/std-js/Notification.js',
+		'https://cdn.kernvalley.us/js/std-js/webShareApi.js',
+		'https://cdn.kernvalley.us/js/std-js/esQuery.js',
+		'https://cdn.kernvalley.us/js/std-js/functions.js',
+		'https://cdn.kernvalley.us/components/login-button.js',
+		'https://cdn.kernvalley.us/components/logout-button.js',
+		'https://cdn.kernvalley.us/components/register-button.js',
+		'https://cdn.kernvalley.us/components/gravatar-img.js',
+		'https://cdn.kernvalley.us/js/std-js/asyncDialog.js',
+		'https://cdn.kernvalley.us/js/User.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentRequest.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentRequestUpdateEvent.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentAddress.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/PaymentResponse.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/BasicCardResponse.js',
+		'https://cdn.kernvalley.us/js/PaymentAPI/BillingAddress.js',
+		'https://cdn.kernvalley.us/components/payment-form/payment-form.js',
+		'https://cdn.kernvalley.us/components/login-form/login-form.js',
+		'https://cdn.kernvalley.us/components/registration-form/registration-form.js',
+		'https://cdn.kernvalley.us/components/login-form/login-form.html',
+		'https://cdn.kernvalley.us/components/registration-form/registration-form.html',
+		'/css/index.css',
+		'/css/vars.css',
+		'/css/layout.css',
+		'/css/header.css',
+		'/css/nav.css',
+		'/css/main.css',
+		'/css/sidebar.css',
+		'/css/footer.css',
+		'https://cdn.kernvalley.us/css/core-css/rem.css',
+		'https://cdn.kernvalley.us/css/core-css/viewport.css',
+		'https://cdn.kernvalley.us/css/core-css/element.css',
+		'https://cdn.kernvalley.us/css/core-css/class-rules.css',
+		'https://cdn.kernvalley.us/css/core-css/utility.css',
+		'https://cdn.kernvalley.us/css/core-css/fonts.css',
+		'https://cdn.kernvalley.us/css/core-css/animations.css',
+		'https://cdn.kernvalley.us/css/normalize/normalize.css',
+		'https://cdn.kernvalley.us/css/animate.css/animate.css',
+		'https://cdn.kernvalley.us/css/core-css/theme/default/index.css',
+		'https://cdn.kernvalley.us/css/core-css/theme/default/dark.css',
+		'https://cdn.kernvalley.us/css/core-css/theme/default/light.css',
+		'https://cdn.kernvalley.us/css/core-css/theme/base.css',
+		'https://cdn.kernvalley.us/css/core-css/layout/shared.css',
+		'/img/apple-touch-icon.png',
+		'/img/icon-192.png',
+		'/img/favicon.svg',
+		'/img/octicons/mail.svg',
+		'/img/logos/instagram.svg',
+		'/img/icons.svg',
+		'/img/listings/05c6cbb6-713f-486d-a867-e3f70d967bb5.jpg',
+		'https://cdn.kernvalley.us/fonts/roboto.woff2',
+		'https://cdn.kernvalley.us/fonts/Libertine.woff',
+		'/listings.json',
+	].map(path => new URL(path, location.origin).href),
+};
